Fixes as suggested in the comments for pull request #1238
