+++ nbgrader/exchange/abc/collect.py
@@ -14,13 +14,3 @@ class ExchangeCollect(Exchange):
         default_value=True,
         help="Whether to cross-check the student_id with the UNIX-owner of the submitted directory."
     ).tag(config=True)
-
-
-    def init_src(self):
-        pass
-
-    def init_dest(self):
-        pass
-
-    def copy_files(self):
-        pass
