+++ nbgrader/exchange/abc/list.py
@@ -1,14 +1,5 @@
-
-import hashlib
-
 from traitlets import Bool
 from .exchange import Exchange
-
-
-def _checksum(path):
-    m = hashlib.md5()
-    m.update(open(path, 'rb').read())
-    return m.hexdigest()
 
 
 class ExchangeList(Exchange):
@@ -17,20 +8,21 @@ class ExchangeList(Exchange):
     cached = Bool(False, help="List assignments in submission cache.").tag(config=True)
     remove = Bool(False, help="Remove, rather than list files.").tag(config=True)
 
-    def parse_assignment(self, assignment):
-        pass
-
-    def format_inbound_assignment(self, info):
-        pass
-
-    def format_outbound_assignment(self, info):
-        pass
-
-    def parse_assignments(self):
-        pass
-
     def list_files(self):
-        pass
+        """Return list of available files """
+        raise NotImplementedError
 
     def remove_files(self):
-        pass
+        """Remove available files """
+        raise NotImplementedError
+
+    def start(self):
+        if self.inbound and self.cached:
+            self.fail("Options --inbound and --cached are incompatible.")
+
+        super(ExchangeList, self).start()
+
+        if self.remove:
+            return self.remove_files()
+        else:
+            return self.list_files()
