+++ nbgrader/exchange/default/__init__.py
@@ -1,10 +1,8 @@
 from .exchange import ExchangeError, Exchange
 from .submit import ExchangeSubmit
 from .release_feedback import ExchangeReleaseFeedback
-from .release import ExchangeRelease
 from .release_assignment import ExchangeReleaseAssignment
 from .fetch_feedback import ExchangeFetchFeedback
-from .fetch import ExchangeFetch
 from .fetch_assignment import ExchangeFetchAssignment
 from .collect import ExchangeCollect
 from .list import ExchangeList
@@ -13,11 +11,9 @@ __all__ = [
     "Exchange",
     "ExchangeError",
     "ExchangeCollect",
-    "ExchangeFetch",
     "ExchangeFetchAssignment",
     "ExchangeFetchFeedback",
     "ExchangeList",
-    "ExchangeRelease",
     "ExchangeReleaseAssignment",
     "ExchangeReleaseFeedback",
     "ExchangeSubmit"
